move signed header, commit, and validator set impls into separate module
