+++ tendermint/src/lite/verifier.rs
@@ -8,13 +8,14 @@
 //! // looks using the types and methods in this crate/module.
 //! ```
 
+use std::cmp::Ordering;
+use std::time::{Duration, SystemTime};
+
 use crate::block::Height;
 use crate::lite::{
     Commit, Error, Header, Requester, SignedHeader, Store, TrustThreshold, TrustedState,
     ValidatorSet,
 };
-use std::cmp::Ordering;
-use std::time::{Duration, SystemTime};
 
 /// Returns an error if the header has expired according to the given
 /// trusting_period and current time. If so, the verifier must be reset subjectively.
@@ -63,7 +64,7 @@ where
     }
 
     // ensure the validator size matches the commit size
-    // NOTE: this is commit structure specifc and should be
+    // NOTE: this is commit structure specific and should be
     // hidden from the light client ...
     if vals.len() != commit.votes_len() {
         return Err(Error::InvalidCommitLength);
@@ -343,10 +344,12 @@ where
 }
 
 mod tests {
-    use super::*;
-    use crate::{hash::Algorithm, Hash};
     use serde::Serialize;
     use sha2::{Digest, Sha256};
+
+    use crate::{hash::Algorithm, Hash};
+
+    use super::*;
 
     #[derive(Clone, Debug, Serialize)]
     struct MockHeader {
