+++ tendermint/src/lite_impl.rs
@@ -1,3 +1,5 @@
 //! Implementation of all light client traits.
 
 pub mod header;
+pub mod signed_header;
+pub mod validator_set;
